Add ClimaCoreRecipesBaseExt extension
