--- .github/workflows/ClimaCorePlots.yml
@@ -1,48 +1,0 @@
-name: ClimaCorePlots CI
-on:
-  pull_request:
-
-# Needed to allow julia-actions/cache to delete old caches that it has created
-permissions:
-  actions: write
-  contents: read
-
-concurrency:
-  group: ${{ github.workflow }}-${{ github.ref }}
-  cancel-in-progress: true
-
-jobs:
-  lib-climacore-plots:
-    runs-on: ubuntu-24.04
-    timeout-minutes: 30
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v7
-      - uses: julia-actions/setup-julia@v3
-        with:
-          version: '1.10'
-      - uses: julia-actions/cache@v3
-      - name: Install Julia dependencies
-        run: >
-          julia --project=monorepo -e 'using Pkg;
-          Pkg.develop([
-            PackageSpec(path="$(pwd())"),
-            PackageSpec(path="$(pwd())/lib/ClimaCorePlots"),
-          ])'
-      - name: Run the tests
-        continue-on-error: true
-        env:
-          GKSwstype: nul
-          CI_OUTPUT_DIR: output
-        run: >
-          julia --project=monorepo -e 'using Pkg; Pkg.test("ClimaCorePlots")'
-          && echo "TESTS_SUCCESSFUL=true" >> $GITHUB_ENV
-      - name: Upload test Artifacts
-        uses: actions/upload-artifact@v7
-        with:
-          name: climacore-plots-output
-          path: |
-            lib/ClimaCorePlots/test/output
-      - name: Fail after artifacts if tests failed
-        if: ${{ env.TESTS_SUCCESSFUL != 'true' }}
-        run: exit 1
