--- docs/src/lib/ClimaCorePlots.md
@@ -1,7 +1,0 @@
-# ClimaCorePlots.jl
-
-```@meta
-CurrentModule = ClimaCorePlots
-```
-
-ClimaCorePlots.jl provides functionality for plotting ClimaCore fields extending the [Plots.jl](https://github.com/JuliaPlots/Plots.jl) package.
