+++ examples/column/ekman.jl
@@ -145,7 +145,7 @@ using Test
 end
 
 ENV["GKSwstype"] = "nul"
-using ClimaCorePlots, Plots
+using Plots
 Plots.GRBackend()
 
 dir = "ekman"
