+++ examples/column/heat.jl
@@ -108,7 +108,7 @@ using Test
 end
 
 ENV["GKSwstype"] = "nul"
-using ClimaCorePlots, Plots
+using Plots
 Plots.GRBackend()
 
 dir = "heat"
