+++ examples/sphere/limiters_advection.jl
@@ -63,7 +63,7 @@ const lim_flag = true
 
 # Plot variables and auxiliary function
 ENV["GKSwstype"] = "nul"
-import ClimaCorePlots, Plots
+import Plots
 Plots.GRBackend()
 dirname = "cg_sphere_advection_limiter_$(test_name)"
