+++ ext/ClimaCoreRecipesBaseExt.jl
@@ -1,0 +1,22 @@
+module ClimaCoreRecipesBaseExt
+
+import RecipesBase
+
+import ClimaComms
+
+import ClimaCore:
+    ClimaCore,
+    DataLayouts,
+    Fields,
+    Geometry,
+    Meshes,
+    Operators,
+    Quadratures,
+    Topologies,
+    Spaces,
+    Utilities
+
+include("plots/tripcolor.jl")
+include("plots/recipes.jl")
+
+end
