+++ lib/ClimaCorePlots/Project.toml
@@ -1,19 +1,13 @@
 name = "ClimaCorePlots"
 uuid = "cf7c7e5a-b407-4c48-9047-11a94a308626"
 authors = ["CliMA Contributors <clima-software@caltech.edu>"]
-version = "0.2.12"
+version = "0.3.0"
 
 [deps]
-ClimaComms = "3a4d1b5c-c61d-41fd-a00a-5873ba7a1b0d"
 ClimaCore = "d414da3d-4745-48bb-8d80-42e94e092884"
-RecipesBase = "3cdcf5f2-1ef4-517c-9805-6587b60abb01"
-StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
-TriplotBase = "981d1d27-644d-49a2-9326-4793e63143c3"
+Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 
 [compat]
-ClimaComms = "0.6"
-ClimaCore = "0.11, 0.12, 0.13, 0.14, 0.15"
-RecipesBase = "1"
-StaticArrays = "1"
-TriplotBase = "0.1"
-julia = "1.7"
+ClimaCore = "0.16"
+Plots = "1"
+julia = "1.10"
