--- lib/ClimaCorePlots/test/Project.toml
@@ -1,10 +1,0 @@
-[deps]
-ClimaComms = "3a4d1b5c-c61d-41fd-a00a-5873ba7a1b0d"
-ClimaCore = "d414da3d-4745-48bb-8d80-42e94e092884"
-ClimaCorePlots = "cf7c7e5a-b407-4c48-9047-11a94a308626"
-IntervalSets = "8197267c-284f-5f27-9208-e0e47529a953"
-Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-
-[compat]
-Plots = "1"
