+++ lib/README.md
@@ -1,6 +1,6 @@
 ## Packages that extend ClimaCore functionality
 
-* [ClimaCoreMakie](https://github.com/CliMA/ClimaCore.jl/tree/main/lib/ClimaCoreMakie): ClimaCore [Makie](https://makie.juliaplots.org/stable) 3D & 2D plotting
-* [ClimaCorePlots](https://github.com/CliMA/ClimaCore.jl/tree/main/lib/ClimaCorePlots): ClimaCore [Plots](https://docs.juliaplots.org/latest/) 2D plotting
+* [ClimaCoreMakie](https://github.com/CliMA/ClimaCore.jl/tree/main/lib/ClimaCoreMakie): deprecated shim; the [Makie](https://makie.juliaplots.org/stable) recipes now live in the `ClimaCoreMakieExt` extension of ClimaCore
+* [ClimaCorePlots](https://github.com/CliMA/ClimaCore.jl/tree/main/lib/ClimaCorePlots): deprecated shim; the [Plots](https://docs.juliaplots.org/latest/) recipes now live in the `ClimaCoreRecipesBaseExt` extension of ClimaCore
 * [ClimaCoreTempestRemap](https://github.com/CliMA/ClimaCore.jl/tree/main/lib/ClimaCoreTempestRemap): Interface for using [TempestRemap](https://github.com/ClimateGlobalChange/tempestremap/) with ClimaCore.
 * [ClimaCoreSpectra](https://github.com/CliMA/ClimaCore.jl/tree/main/lib/ClimaCoreSpectra): ClimaCore for spherical harmonic spectra
