+++ test/runtests.jl
@@ -177,13 +177,14 @@ unit_tests = [
     UnitTest("Integration - Solid-body rotation 3D"     ,"Integration/smoke_solid_body_rotation.jl"; meta = :cpu_only, tier = :smoke, subsystem = :integration),
     UnitTest("Integration - column FCT/van Leer advection", "Integration/smoke_column_advection.jl"; meta = :cpu_only, tier = :smoke, subsystem = :limiters),
 
-    # Visualization
-    UnitTest("Makie extension"                          ,"Visualize/visualize.jl"; meta = :cpu_only, tier = :unit, subsystem = :other),
-
     # Quality & Deprecations
     UnitTest("Aqua"                                     ,"aqua.jl"; tier = :misc, subsystem = :quality),
     UnitTest("Deprecations"                             ,"deprecations.jl"; tier = :misc, subsystem = :quality),
     UnitTest("Precompile workload"                      ,"precompile_workload.jl"; tier = :misc, subsystem = :quality, slow = true), # ~1 min: recompiles ClimaCore in a subprocess
+
+    # Visualization
+    UnitTest("Makie extension"                          ,"Visualize/visualize.jl"; meta = :cpu_only, tier = :unit, subsystem = :viz),
+    UnitTest("Plots extension"                          ,"Visualize/plots.jl"; meta = :cpu_only, tier = :unit, subsystem = :viz),
 
     # GPU Only
     UnitTest("GPU - cuda"                               ,"gpu/cuda.jl"; meta = :gpu_only, tier = :gpu, subsystem = :gpu),
